Started mySet.cpp
I need to brush up on how header and object files work, but the dream here is to implement my own Set class in C++ using my own hash function.

Note: The hash function is not perfect. it is inspired by one found here: https://algs4.cs.princeton.edu/34hash/
with a twist, namely that we *square* the number associated w/ each letter in the string.